django-dashboard-kit/requirements.txt
asgiref==3.8.1
autopep8==1.6.0
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
dj-database-url==0.5.0
Django==5.0.7
django-crispy-forms==2.2
django-jazzmin==3.0.0
djangorestframework==3.15.2
gunicorn==20.1.0
idna==3.7
pillow==10.4.0
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
requests==2.32.3
setuptools==70.2.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2024.1
Unipath==1.1
urllib3==2.2.2
whitenoise==5.3.0
